+++ mammoth.repos
@@ -7,23 +7,11 @@ repositories:
     type: git
     url: https://github.com/dcutting133/lidar_processor.git
     version: master
-  isc_roboteq:
+  ros_ign:
     type: git
-    url: https://github.com/iscumd/isc_roboteq_hdc2460.git
+    url: https://github.com/ignitionrobotics/ros_ign.git
     version: ros2
   pointcloud_to_laserscan:
     type: git
     url: https://github.com/ros-perception/pointcloud_to_laserscan.git
     version: foxy
-  serial_ros2:
-    type: git
-    url: https://github.com/iscumd/isc_serial.git
-    version: ros2
-  serial_utils_ros2:
-    type: git
-    url: https://github.com/iscumd/isc_serial_utils.git
-    version: ros2     
-  ros2_ouster:
-    type: git
-    url: https://github.com/ros-drivers/ros2_ouster_drivers.git
-    version: foxy-devel
