+++ mammoth_gazebo/package.xml
@@ -16,14 +16,13 @@
     <buildtool_depend>ament_cmake</buildtool_depend>
 
     <exec_depend>mammoth_description</exec_depend>
+    
     <exec_depend>joint_state_publisher</exec_depend>
     <exec_depend>robot_state_publisher</exec_depend>
     <exec_depend>lidar_processor</exec_depend>
-    <exec_depend>slam_toolbox</exec_depend>
-    <exec_depend>navigation2</exec_depend>
-    <exec_depend>nav2-bringup</exec_depend>
     <exec_depend>ros_ign_bridge</exec_depend>
     <exec_depend>ros_ign_gazebo</exec_depend>
+    <exec_depend>robot_localization</exec_depend>
     <exec_depend>teleop_twist_joy</exec_depend>
     <exec_depend>joy</exec_depend>
