+++ mammoth_gazebo/worlds/test.sdf
@@ -71,10 +71,12 @@
         </visual>
       </link>
     </model>
+    <!--
     <include>
     <uri>
-      https://fuel.ignitionrobotics.org/1.0/OpenRobotics/models/Cafe
+     https://fuel.ignitionrobotics.org/1.0/OpenRobotics/models/Depot
     </uri>
     </include>
+    -->
   </world>
 </sdf>
