--- mammoth_snowplow/config/joystick/switch-pro.config.yaml
@@ -1,18 +1,0 @@
-teleop_twist_joy_node:
-  ros__parameters:
-    axis_linear:  # Left thumb stick vertical
-      x: 1
-    scale_linear:
-      x: 2.0
-    scale_linear_turbo:
-      x: 3.0
-
-    axis_angular:  # Left thumb stick horizontal
-      yaw: 5
-    scale_angular:
-      yaw: 2.0
-    scale_angular_turbo:
-      yaw: 3.0
-
-    enable_button: 2  # Left trigger button
-    enable_turbo_button: 5  # Right trigger button
