--- mammoth_snowplow/config/maps/amcl/cafe.yaml
@@ -1,7 +1,0 @@
-image: cafe.pgm
-mode: trinary
-resolution: 0.03
-origin: [-5.05, -11, 0]
-negate: 0
-occupied_thresh: 0.65
-free_thresh: 0.25
