--- mammoth_snowplow/config/sensors/lidar/pointcloud_to_laserscan.yaml
@@ -1,14 +1,0 @@
-pointcloud_to_laserscan:
-  ros__parameters:
-    target_frame: 'laser_link'
-    transform_tolerance: 0.01
-    min_height: 0.0
-    max_height: 20.0
-    angle_min: -3.1415
-    angle_max:  3.1415
-    angle_increment: 0.0087
-    scan_time: 0.01
-    range_min: 0.45
-    range_max: 35.0
-    use_inf: False
-    inf_epsilon: 1.0
