--- mammoth_snowplow/config/slam/mapper_params_online_async.yaml
@@ -1,71 +1,0 @@
-slam_toolbox:
-  ros__parameters:
-
-    # Plugin params
-    solver_plugin: solver_plugins::CeresSolver
-    ceres_linear_solver: SPARSE_NORMAL_CHOLESKY
-    ceres_preconditioner: SCHUR_JACOBI
-    ceres_trust_strategy: LEVENBERG_MARQUARDT
-    ceres_dogleg_type: TRADITIONAL_DOGLEG
-    ceres_loss_function: None
-
-    # ROS Parameters
-    odom_frame: odom
-    map_frame: map
-    base_frame: base_footprint
-    scan_topic: /scan
-    mode: localization
-    # if you'd like to immediately start continuing a map at a given pose
-    # or at the dock, but they are mutually exclusive, if pose is given
-    # will use pose
-    map_file_name: install/mammoth_snowplow/share/mammoth_snowplow/config/maps/slam_toolbox/cafe
-    map_start_pose: [0.0, 0.0, 0.0]
-    #map_start_at_dock: true
-
-    debug_logging: false
-    throttle_scans: 1
-    transform_publish_period: 0.0375 #if 0 never publishes odometry
-    map_update_interval: 5.0
-    resolution: 0.03
-    max_laser_range: 20.0 #for rastering images
-    minimum_time_interval: 0.5
-    transform_timeout: 0.3
-    tf_buffer_duration: 80.
-    stack_size_to_use: 400000000 #// program needs a larger stack size to serialize large maps
-    enable_interactive_mode: true
-
-    # General Parameters
-    use_scan_matching: true
-    use_scan_barycenter: true
-    minimum_travel_distance: 0.01
-    minimum_travel_heading: 0.01
-    scan_buffer_size: 90
-    scan_buffer_maximum_scan_distance: 10.0
-    link_match_minimum_response_fine: 0.1  
-    link_scan_maximum_distance: 3.0
-    loop_search_maximum_distance: 4.0
-    do_loop_closing: true 
-    loop_match_minimum_chain_size: 20     
-    loop_match_maximum_variance_coarse: 3.0  
-    loop_match_minimum_response_coarse: 0.35    
-    loop_match_minimum_response_fine: 0.45
-
-    # Correlation Parameters - Correlation Parameters
-    correlation_search_space_dimension: 0.8
-    correlation_search_space_resolution: 0.01
-    correlation_search_space_smear_deviation: 0.1 
-
-    # Correlation Parameters - Loop Closure Parameters
-    loop_search_space_dimension: 10.0
-    loop_search_space_resolution: 0.05
-    loop_search_space_smear_deviation: 0.03
-
-    # Scan Matcher Parameters
-    distance_variance_penalty: 0.5      
-    angle_variance_penalty: 1.0    
-    fine_search_angle_offset: 0.00349     
-    coarse_search_angle_offset: 0.349   
-    coarse_angle_resolution: 0.0349        
-    minimum_angle_penalty: 0.9
-    minimum_distance_penalty: 0.5
-    use_response_expansion: true
