--- mammoth_snowplow/launch/include/lidar_processor/lidar_processor.launch.py
@@ -1,68 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2021 Intelligent Systems Club
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-import os
-
-from ament_index_python.packages import get_package_share_directory
-
-from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument
-from launch.actions import IncludeLaunchDescription
-from launch.conditions import IfCondition
-from launch.launch_description_sources import PythonLaunchDescriptionSource
-from launch.substitutions import LaunchConfiguration
-
-from launch_ros.actions import Node
-
-def generate_launch_description():
-    # ROS packages
-    pkg_mammoth_gazebo = get_package_share_directory('mammoth_snowplow')
-
-    # Launch arguments
-    use_sim_time = LaunchConfiguration('use_sim_time', default='false')
-
-    lidar_processor = Node(
-        package='lidar_processor',
-        executable='lidar_processor',
-        name='lidar_processor',
-        output='screen',
-        remappings=[
-            ('/lidar/raw_points', '/mammoth/raw_points'),
-            ('/lidar/filtered_points', '/mammoth/filtered_points'),
-            ('/lidar/unfiltered_points', '/mammoth/unfiltered_points'),
-            ('/lidar/raw_scan', '/mammoth/raw_scan'),
-            ('/lidar/filtered_scan', '/mammoth/filtered_scan'),
-            ('/lidar/unfiltered_scan', '/mammoth/unfiltered_scan'),
-        ],
-        parameters=[{
-            'use_sim_time': use_sim_time
-        }]
-    )
-
-    return LaunchDescription([
-        # Launch Arguments
-        DeclareLaunchArgument('use_sim_time', default_value='false',
-                              description='Use simulation clock if true'),
-
-        # Nodes
-        lidar_processor,
-    ])
