--- mammoth_snowplow/package.xml
@@ -1,25 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>mammoth_snowplow</name>
-  <version>0.0.0</version>
-  <description>TODO: Package description</description>
-  <maintainer email="agirven@umich.edu">avery</maintainer>
-  <license>TODO: License declaration</license>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-
-  <exec_depend>mammoth_description</exec_depend>
-  <exec_depend>ros2_ouster</exec_depend>  
-  <exec_depend>lidar_processor</exec_depend> 
-  <exec_depend>isc_roboteq</exec_depend>     
-  <exec_depend>realsense2_camera</exec_depend>   
-  <test_depend>ament_lint_auto</test_depend>
-
-  <test_depend>ament_lint_common</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
-
